requirements.txt
Pillow>=9.2.0
numpy>=1.21.0
torch>=1.12.0